구조 변경 및 파일정리
