+++ Capstone Design_Miracle Energy/EPS/Nano/AI/neural_training.py
