--- Capstone Design_Miracle Energy/EPS/Nano/data/test.csv
@@ -1,165 +1,0 @@
-211
-197
-192
-196
-185
-204
-196
-194
-182
-198
-192
-172
-169
-177
-191
-181
-175
-177
-190
-186
-183
-190
-195
-185
-194
-177
-191
-178
-270
-236
-200
-190
-173
-176
-187
-176
-168
-196
-177
-181
-238
-233
-196
-182
-179
-178
-183
-290
-246
-208
-175
-191
-263
-223
-194
-190
-182
-181
-188
-191
-186
-187
-175
-179
-191
-249
-210
-206
-171
-176
-175
-195
-180
-186
-192
-186
-201
-248
-210
-189
-176
-181
-183
-173
-173
-154
-200
-182
-209
-245
-208
-239
-201
-176
-176
-176
-193
-193
-184
-178
-182
-261
-221
-185
-175
-186
-188
-184
-190
-184
-198
-190
-198
-254
-238
-204
-195
-183
-175
-179
-177
-175
-187
-183
-178
-277
-275
-234
-235
-276
-236
-251
-238
-202
-213
-254
-215
-284
-267
-226
-292
-268
-250
-211
-265
-270
-230
-193
-246
-249
-271
-305
-292
-250
-221
-223
-188
-184
-252
-218
-183
-251
-272
-245
-206
