--- Capstone Design_Miracle Energy/EPS/comsume_serial.py
@@ -1,208 +1,0 @@
-#include <avr/wdt.h>
-#include <Filters.h> //Easy library to do the calculations
-#include <Wire.h>
-
-def testFrequency =60;                      // test signal frequency (Hz):
-def windowLength = 0.0/testFrequenncy;     // how long to average the signal, for statistist
-:
-
-def Sensor = 0; //Snsor analog inpput, here it's A0
-: to average the signal, for statistist
-:
-
-def intercept = -0.4; // to be adjjusted based on calibration testing
-:nal, for statistist
-:
-def slope = 0.0405;// to be adjustted based on calibration testing
-:
-:nal, for statistist
-:
-def current_Volts; / Voltage
-:stteed based on calibration testing
-:
-:nal, for statistist
-:
-def voltajeSensor;
-Voltage
-:stteedd based on calibration testing
-:
-:nal, for statistist
-:
-def corriente=0;
-:Vltage
-:stteedd  based on calibration testing
-:
-:nal, for statistis
-
-:
-def Sumatoria=0;
-:lage
-:stteedd  bbased on calibration testing
-:
-:nal, for statistis
-::
-long tiempo=millis(); 
-def N=0;
-:ia=0;
-:lae
-:stteedd  bbaased on calibration testing
-:
-:nal, for statistis
-::
-unsigned long printPeriod = 1000; //Refresh rate
-unsigned long previousMillis = 0;
-unsigned long previousMillis2 = 0;
-const unsigned long intervalo=0;
-
-def get_corriente():                                                                 
-{
-  float voltajeSensor;
-  float corriente=0;
-  float Sumatoria=0;
-  long tiempo=millis(); 
-  int N=0;
-  while(millis()-                                                              
-  { 
-    voltajeSensor = analogRead(A0)*                                                          
-    corriente=voltajeSensor*70.5; //corriente=VoltajeSensor*(100A/1V)
-    Sumatoria=Sumatoria+sq(corriente);                         
-    N=N+1;
-    delay(1);   
-  }
-  Sumatoria=Sumatoria*2;//Para compensar los cuadrados de los semiciclos negativos.
-  corriente=sqrt((Sumatoria)/                        
-  corriente=corriente;
-  return(corriente); 
-}
-
-def get_corriente2(:teedd  bbaasseed on calibration testing
-:
-:nal, for statistis
-::
-{
-  float voltajeSensor2;
-  float corriente2=0;
-  float Sumatoria2=0;
-  long tiempo2=millis(); 
-  int N2=0;
-  while(millis()-                                                               
-  { 
-    voltajeSensor2 = analogRead(A2)*                                                          
-    corriente2=voltajeSensor2*187.0; //corriente=VoltajeSensor*(100A/1V)
-    Sumatoria2=Sumatoria2+sq(corriente2);                         
-    N2=N2+1;
-    delay(1);   
-  }
-  Sumatoria2=Sumatoria2*2;//Para compensar los cuadrados de los semiciclos negativos.
-  corriente2=sqrt((Sumatoria2)/                         
-  corriente2=corriente2;
-  return(corriente2); 
-}
-
-def get_corriente3(:edd  bbaasseedd on calibration testing
-:
-:nal, for statistis
-::
-{
-  float voltajeSensor3;
-  float corriente3=0;
-  float Sumatoria3=0;
-  long tiempo3=millis(); 
-  int N3=0;
-  while(millis()-                                                               
-  { 
-    voltajeSensor3 = analogRead(A3)*                                                          
-    corriente3=voltajeSensor3*157.0; //corriente=VoltajeSensor*(100A/1V)
-    Sumatoria3=Sumatoria3+sq(corriente3);                         
-    N3=N3+1;
-    delay(1);   
-  }
-  Sumatoria3=Sumatoria3*2;//Para compensar los cuadrados de los semiciclos negativos.
-  corriente3=sqrt((Sumatoria3)/                         
-  corriente3=corriente3;
-  return(corriente3); 
-}
-def get_corriente4(:d  bbaasseedd  on calibration testing
-:
-:nal, for statistis
-::
-{
-  float voltajeSensor4;
-  float corriente4=0;
-  float Sumatoria4=0;
-  long tiempo4=millis(); 
-  int N4=0;
-  while(millis()-                                                               
-  { 
-    voltajeSensor4 = analogRead(A4)*                                                          
-    corriente4=voltajeSensor4*127.0; //corriente=VoltajeSensor*(100A/1V)
-    Sumatoria4=Sumatoria4+sq(corriente4);                         
-    N4=N4+1;
-    delay(1);   
-  }
-  Sumatoria4=Sumatoria4*2;//Para compensar los cuadrados de los semiciclos negativos.
-  corriente4=sqrt((Sumatoria4)/                         
-  corriente4=corriente4;
-  return(corriente4); 
-}
-
-def setup():                                                                     
- Serial.begin(9600);  
- pinMode(7, OUTPUT);                              
- wdt_enable(WDTO_8S); 
-}
-
-def loop():                                                                     
-unsigned long ahora=millis(); 
-if (ahora - previousMillis2 >= intervalo):
- float I1 = get_corriente(); 
- float I2 = get_corriente2(); 
- float I3 = get_corriente3(); 
- float I4 = get_corriente4(); 
- float Itot = get_corriente2()                     
-   
-digitalWrite(7, LOW); 
-
-RunningStatistics inputStats;                //Easy life lines, actual calculation of the RMS requirinputStats.setWindowSecs( windowLength ):
-inputStats.setWindowSecs( windowLength ):
-def var=0; 
-:e4(: baasseedd  oonn  calibration testing
-:
-:nal, for statistis
-::
-while( var<3 )        
-    Sensor = analogRead(A5);                               
-    inputStats.input(Sensor);                           
-     
-    if((unsigned long):
-      previousMillis = millis();                               
-      current_Volts = intercept + slope * inputStats.sigma();                                                                                                                                                                        
-                                                                                                                                 
-                             
-      var+=1;
-      
-      }
-      else:
-      var+=1;}
-      
-      if (current_Volts<=1):
-        current_Volts=0;}
-      }
-    }
-  // Uploads new telemetry to ThingsBoard using MQTT.
-  float PCONS = I1*current_Volts;
-  digitalWrite(7, HIGH); 
-  Serial.print(PCONS); 
-  Serial.print("x"); 
-  Serial.print(current_Volts); 
-  Serial.print("x"); 
-  Serial.println(I1); 
-
-digitalWrite(7, LOW);  
-delay(1000); 
-
-previousMillis2=ahora;}
-wdt_reset(); 
-
-    }
-
