+++ Capstone Design_Miracle Energy/esp8266/readme.txt
@@ -1,0 +1,3 @@
+IoT 센서 참고사항
+1. 툴-> 라이브러리에 PubSubClient를 입력하여 다운로드 받을것.
+2. Nano에는 모스키토, 모스키토-클라이언트가 설치되어있을것.(모스키토는 브로커역할을 함)
