+++ Capstone Design_Miracle Energy/결과이미지/predict.png
